split mat_znx into struct and ops + added missing ops on module
